+++ frontend-web/src/main.css
@@ -10,56 +10,3 @@ main ul{
     list-style: none;
 }
 
-.dev-item{
-    background: #ffffff;
-    box-shadow: 0 0 14px 0 rgba(0, 0, 0, 0.02);
-    border-radius: 2px;
-    padding: 20px;
-}
-
-.dev-item header {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-}
-
-.dev-item header img {
-    width: 54px;
-    height: 54px;
-    border-radius: 50%;
-}
-
-.dev-item header .user-info {
-    margin-left: 10px;
-}
-
-.user-info strong{
-    display: block;
-    font-size: 16px;
-    color: #333;
-}
-
-.user-info span{
-    font-size: 13px;
-    color: #999;
-    margin-top: 2px;
-}
-
-.dev-item p{
-    color: #666;
-    font-size: 14px;
-    line-height: 20px;
-    margin: 10px;
-}
-
-.dev-item a{
-    color: #8e4dff;
-    font-size: 15px;
-    text-decoration: none;
-    transition: color 0.5s;
-}
-
-.dev-item a:hover{
-    color: #5a2ea6;
-    
-}
